refactor to separate TVM specific bits from paged_cache_model
